Accounting for network problems causing tar not to install. Skipping and leaving the tarball behind if tar isn't available and messaging the user.

+++ mkvm.firstboot.sh
@@ -19,21 +19,26 @@ service sshd start
 # When the network is configured, you can install stuff
 yum install tar -y --nogpgcheck
 
-# Run plugins
-tar xzf /root/plugins.tar.gz -C /root/
-for plugin in /root/plugins/*; do 
-    echo "Starting plugin '$(basename ${plugin})' install..."
-    if [ -e ${plugin}/init.sh ]; then
-        chmod +x ${plugin}/init.sh # Make sure it's executable.
-        PLUGINS_DIR=${plugin} sh ${plugin}/init.sh
-        echo "Plugin '$(basename ${plugin})' installation complete!"
-    else
-        echo "Plugin '$(basename ${plugin})' has a missing or disabled init.sh. Skipping plugin."
-    fi
-done
+# check exit status of tar installation
+if [ $? -ne 0 ]; then
+    echo "The tar package failed to install. Skipping plugins and leaving plugins.tar.gz for later use."
+else
+    # Run plugins
+    tar xzf /root/plugins.tar.gz -C /root/
+    for plugin in /root/plugins/*; do 
+        echo "Starting plugin '$(basename ${plugin})' install..."
+        if [ -e ${plugin}/init.sh ]; then
+            chmod +x ${plugin}/init.sh # Make sure it's executable.
+            PLUGINS_DIR=${plugin} sh ${plugin}/init.sh
+            echo "Plugin '$(basename ${plugin})' installation complete!"
+        else
+            echo "Plugin '$(basename ${plugin})' has a missing or disabled init.sh. Skipping plugin."
+        fi
+    done
 
-# Cleanup plugins tar
-rm -f /root/plugins.tar.gz
+    # Cleanup plugins tar
+    rm -f /root/plugins.tar.gz
+fi
 
 echo "==========================="
 echo "VM configuration complete!!"
